Move Hamcrest Mockito matchers to internal testing
