+++ subprojects/internal-testing/src/main/java/dev/nokee/internal/testing/invocations/HasInvocationResults.java
@@ -13,8 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.xcode.project.invocations;
+package dev.nokee.internal.testing.invocations;
 
-public interface InvocationResult0 extends InvocationResult {
+import java.util.List;
 
+public interface HasInvocationResults<A extends InvocationResult> {
+	List<A> getAllInvocations();
 }
